♻️ Add SchemaPage with ERDEditor, safe override utility, constants
