+++ frontend/apps/app/features/schemas/pages/SchemaPage/components/ERDEditor/index.ts
@@ -1,0 +1,1 @@
+export * from './ERDEditor'
