+++ frontend/apps/app/features/schemas/pages/SchemaPage/index.ts
@@ -1,0 +1,1 @@
+export * from './SchemaPage'
